@@ -9,30 +9,32 @@
   "dependencies": {
     "axios": "^0.19.0",
     "core-js": "^3.4.4",
-    "vue": "^2.6.10",
-    "vue-router": "^3.1.3",
-    "vuex": "^3.1.2",
-    "prismjs": "^1.21.0"
+    "mitt": "^2.1.0",
+    "vue": "^3.0.2",
+    "vue-router": "^4.0.0-beta.13",
+    "vuex": "^4.0.0-beta.4"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.1.0",
-    "@vue/cli-plugin-eslint": "^4.1.0",
-    "@vue/cli-plugin-router": "^4.1.0",
-    "@vue/cli-plugin-typescript": "^4.1.0",
-    "@vue/cli-plugin-vuex": "^4.1.0",
-    "@vue/cli-service": "^4.1.0",
-    "@vue/eslint-config-airbnb": "^4.0.0",
-    "@vue/eslint-config-typescript": "^4.0.0",
-    "babel-plugin-prismjs": "^2.0.1",
+    "@typescript-eslint/eslint-plugin": "^2.33.0",
+    "@typescript-eslint/parser": "^2.33.0",
+    "@vue/cli-plugin-babel": "~4.5.8",
+    "@vue/cli-plugin-eslint": "~4.5.8",
+    "@vue/cli-plugin-router": "~4.5.8",
+    "@vue/cli-plugin-typescript": "~4.5.8",
+    "@vue/cli-plugin-vuex": "~4.5.8",
+    "@vue/cli-service": "~4.5.8",
+    "@vue/compiler-sfc": "^3.0.0-beta.1",
+    "@vue/eslint-config-airbnb": "^5.0.2",
+    "@vue/eslint-config-typescript": "^5.0.2",
+    "eslint": "^6.7.2",
+    "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-vue": "^7.0.0-alpha.0",
     "path": "^0.12.7",
-    "eslint": "^5.16.0",
-    "eslint-plugin-vue": "^5.0.0",
     "sass": "^1.23.7",
     "sass-loader": "^8.0.0",
     "style-resources-loader": "^1.3.3",
-    "typescript": "~3.5.3",
-    "vue-multiselect": "^2.1.6",
-    "vue-template-compiler": "^2.6.10",
-    "vue-cli-plugin-markdown": "^0.1.2"
+    "typescript": "~3.9.3",
+    "vue-cli-plugin-markdown": "^0.1.2",
+    "vue-multiselect": "^2.1.6"
   }
 }
