@@ -1,67 +1,69 @@
 <template>
 	<section class="app-main">
 
-		<!-- GLOBAL SPINNER -->
+		<!-- GLOBAL SPINNER
         <transition name="fade" mode="out-in">
 
             <div v-show="showGlobalSpinner" class="global-spinner-con">
                 <loading></loading>
             </div>
 
-        </transition>
+        </transition>  -->
 
 
-		<!-- APP ACTUAL -->
-        <transition-group name="fade" mode="out-in">
+		<!-- APP ACTUAL
 
-            <template v-show="!showGlobalSpinner">
+        <template v-show="!showGlobalSpinner">  
 
-                <sitemodal :key="1" v-if="showsitemodal" v-bind="{ contains }"></sitemodal>
+        <sitemodal :key="1" v-if="showsitemodal" v-bind="{ contains }"></sitemodal> -->
 
-                <navbar :key="2" v-if="checkroute"></navbar>
+        <navbar :key="2"></navbar>
 
-                <router-view :key="3"></router-view>
+        <router-view :key="3"></router-view>
 
-                <qrfooter :key="4" v-if="loadPushed && showfooter"></qrfooter>
+        <qrfooter :key="4"></qrfooter>
 
-            </template>
-            
-        </transition-group>
 
         <usermessages v-bind="{ msg, black, sass }" v-if="showUserMessage"></usermessages>
 
-        <sitemodal contains="smallscreen" v-if="windowWidth < 620"></sitemodal>
+        <!-- <sitemodal contains="smallscreen" v-if="windowWidth < 620"></sitemodal> -->
         
 	</section>
 </template>
 
-<script>
-import navbar from './components/nav/navbar'
-import qrfooter from './components/nav/qrfooter'
-import usermessages from './components/usermessages'
-import loading from './components/loading'
-import debounce from './utils/functions'
-import sitemodal from './components/sitemodal/sitemodal'
+<script lang="ts">
+
+import { defineComponent } from 'vue'
+
+import navbar from './components/nav/navbar.vue'
+import qrfooter from './components/nav/qrfooter.vue'
+import usermessages from './components/usermessages.vue'
+import loading from './components/loading.vue'
+// import debounce from './utils/functions'
+import sitemodal from './components/sitemodal/sitemodal.vue'
+
 import { mapGetters, mapMutations } from 'vuex'
+
 import { 
     EventBus,
     LOADING,
     MESSAGES,
     SITEMODAL,
     EDITOR_ERROR,
-} from './EventBus.ts'
-import { LOGGED_IN_WITH_GOOGLE } from './EventBus'
+    NEED_TO_BE_LOGGED_IN,
+    LOGGED_IN_WITH_GOOGLE,
+} from './EventBus'
 
 // useful cludge
-let __proxy
+let __proxy: any
 
-export default {
+const App = defineComponent({
     name: 'app',
     components: {
         navbar,
         qrfooter,
-        sitemodal,
-        loading,
+        // sitemodal,
+        // loading,
         usermessages,
     },
     data () {
@@ -86,15 +88,15 @@ export default {
     mounted () {
 
         // EventBus handling global loading spinner, user message pop up and site wide modals
-        EventBus.$on(LOADING, is => { this.showGlobalSpinner = is })
+        EventBus.on(LOADING, is => { this.showGlobalSpinner = is })
 
-        EventBus.$on(MESSAGES, deets => {
+        EventBus.on(MESSAGES, deets => {
             this.showUserMessage = deets.is
             this.msg = deets.msg
             this.sass = deets.color
         })
 
-        EventBus.$on(SITEMODAL, (type = null) => {
+        EventBus.on(SITEMODAL, (type = null) => {
             if (type) {
                 this.contains = type
                 this.showsitemodal = true
@@ -107,22 +109,22 @@ export default {
         const para = new URLSearchParams(window.location.search)
         if (para.get('redirect') === 'true') {
 
-            EventBus.$emit(MESSAGES, NEED_TO_BE_LOGGED_IN)
+            EventBus.emit(MESSAGES, NEED_TO_BE_LOGGED_IN)
 
             this.$router.push({ path: '/auth/login?redirect=true' })
         }
 
         if (para.get('googleauth') === 'true') {
 
-            EventBus.$emit(MESSAGES, LOGGED_IN_WITH_GOOGLE)
+            EventBus.emit(MESSAGES, LOGGED_IN_WITH_GOOGLE)
             this.$router.push({ path: '/app' })
             
         }
 
         // useful cludge
         __proxy = this
-        window.addEventListener('resize', debounce(this.sizeChange, 500))
-        window.addEventListener('scroll', debounce(this.scrollChange, 100))
+        window.addEventListener('resize', this.sizeChange)
+        window.addEventListener('scroll', this.scrollChange)
 
         // another useful cludge to prevent the footer from flashing
         setTimeout(() => this.loadPushed = true, 1500)  
@@ -141,19 +143,21 @@ export default {
 
         },
     },
-    computed: {
-        checkroute() {
-            return ['wapp'].indexOf(this.$route.name) === -1
-        },
-        showfooter() {
-            return ['auth'].indexOf(this.$route.name) === -1
-        },
-        ...mapGetters(['windowWidth']),
-    },
-}
+    // computed: {
+    //     checkroute() {
+    //         return ['wapp'].indexOf(this.$route.name) === -1
+    //     },
+    //     showfooter() {
+    //         return ['auth'].indexOf(this.$route.name) === -1
+    //     },
+    //     ...mapGetters(['windowWidth']),
+    // },
+})
+
+export default App
 </script>
 
-<style lang="sass">
+<style lang="sass" scoped>
 .app-main
     min-height: 100vh !important
 .global-spinner-con
