@@ -1,23 +1,28 @@
 import 'vue-multiselect/dist/vue-multiselect.min.css'
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router'
 import store from './store'
 import { QAuth } from './api/auth'
 import { QAdmin } from './api/admin'
 
-Vue.config.productionTip = false
+const app = createApp(App)
 
-Vue.prototype.$QAdmin = new QAdmin()
+app.config.globalProperties.$QAdmin = new QAdmin()
+
+const qAuth = new QAuth(store)
+app.config.globalProperties.$QAuth = qAuth
+
+export default qAuth
 
 /* eslint-disable no-console */
-Vue.config.errorHandler = (error, vm, info) => {
+app.config.errorHandler = (error, vm, info) => {
 
     if (process.env.NODE_ENV === 'development') {
         console.error(error, vm, info)
     }
 
-    Vue.prototype.$QAdmin.clientsideerror({
+    app.config.globalProperties.$QAdmin.clientsideerror({
         time: new Date(),
         userAgent: navigator.userAgent,
         error, info,
@@ -32,7 +37,7 @@ window.onerror = (message, url, line, column, error) => {
         console.error(message, url, line, column, error)
     }
 
-    Vue.prototype.$QAdmin.clientsideerror({
+    app.config.globalProperties.$QAdmin.clientsideerror({
         time: new Date(),
         userAgent: navigator.userAgent,
         message, url, line, column, error,
@@ -42,12 +47,4 @@ window.onerror = (message, url, line, column, error) => {
 }
 /* eslint-enable no-console */
 
-new Vue({
-    router,
-    store,
-    render: h => h(App),
-}).$mount('#app')
-
-const qAuth = new QAuth(store)
-Vue.prototype.$QAuth = qAuth
-export default qAuth
+app.use(router).use(store).mount('#app')
