@@ -1,14 +1,11 @@
-import Vue from 'vue'
-import Vuex from 'vuex'
+import { createStore } from 'vuex'
 
 import state from './state'
 import getters from './getters'
 import mutations from './mutations'
 import actions from './actions'
 
-Vue.use(Vuex)
-
-export default new Vuex.Store({
+export default createStore({
     state,
     actions,
     getters,
